requirements.txt

streamlit
snowflake-snowpark-python
pandas
plotly
numpy
